Safari-611.1.10/LayoutTests/media/video-fullscreen-only-controls-expected.txt

This tests that when inline-playback is restricted, the video element's "controls" are always shown while inline.


RUN(internals.settings.setAllowsInlineMediaPlayback(false))
EXPECTED (video.hasAttribute('controls') == 'false') OK
EXPECTED (video.controls == 'false') OK
EXPECTED (shadowRoot = internals.shadowRoot(video) != 'null') OK
EXPECTED (panel = mediaControlsElement(shadowRoot.firstChild, '-webkit-media-controls-panel') != 'null') OK
EXPECTED (internals.shadowPseudoId(panel) == '-webkit-media-controls-panel') OK
EXPECTED (document.defaultView.getComputedStyle(panel)['display'] != 'none') OK
EXPECTED (document.defaultView.getComputedStyle(panel)['height'] >= '20px') OK
END OF TEST
